// src/api/models/TeamRecord.ts
/* tslint:disable */
/* eslint-disable */
/**
 * radarvan
 * Stats for generals
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TeamRecord
 */
export interface TeamRecord {
    /**
     * 
     * @type {Array<string>}
     * @memberof TeamRecord
     */
    players: Array<string>;
    /**
     * 
     * @type {number}
     * @memberof TeamRecord
     */
    wins: number;
    /**
     * 
     * @type {number}
     * @memberof TeamRecord
     */
    losses: number;
}

/**
 * Check if a given object implements the TeamRecord interface.
 */
export function instanceOfTeamRecord(value: object): value is TeamRecord {
    if (!('players' in value) || value['players'] === undefined) return false;
    if (!('wins' in value) || value['wins'] === undefined) return false;
    if (!('losses' in value) || value['losses'] === undefined) return false;
    return true;
}

export function TeamRecordFromJSON(json: any): TeamRecord {
    return TeamRecordFromJSONTyped(json, false);
}

export function TeamRecordFromJSONTyped(json: any, ignoreDiscriminator: boolean): TeamRecord {
    if (json == null) {
        return json;
    }
    return {
        
        'players': json['players'],
        'wins': json['wins'],
        'losses': json['losses'],
    };
}

export function TeamRecordToJSON(json: any): TeamRecord {
    return TeamRecordToJSONTyped(json, false);
}

export function TeamRecordToJSONTyped(value?: TeamRecord | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'players': value['players'],
        'wins': value['wins'],
        'losses': value['losses'],
    };
}