// src/api/models/TeamSizeGroup.ts
/* tslint:disable */
/* eslint-disable */
/**
 * radarvan
 * Stats for generals
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TeamRecord } from './TeamRecord';
import {
    TeamRecordFromJSON,
    TeamRecordFromJSONTyped,
    TeamRecordToJSON,
    TeamRecordToJSONTyped,
} from './TeamRecord';

/**
 * 
 * @export
 * @interface TeamSizeGroup
 */
export interface TeamSizeGroup {
    /**
     * 
     * @type {number}
     * @memberof TeamSizeGroup
     */
    size: number;
    /**
     * 
     * @type {Array<TeamRecord>}
     * @memberof TeamSizeGroup
     */
    teams: Array<TeamRecord>;
}

/**
 * Check if a given object implements the TeamSizeGroup interface.
 */
export function instanceOfTeamSizeGroup(value: object): value is TeamSizeGroup {
    if (!('size' in value) || value['size'] === undefined) return false;
    if (!('teams' in value) || value['teams'] === undefined) return false;
    return true;
}

export function TeamSizeGroupFromJSON(json: any): TeamSizeGroup {
    return TeamSizeGroupFromJSONTyped(json, false);
}

export function TeamSizeGroupFromJSONTyped(json: any, ignoreDiscriminator: boolean): TeamSizeGroup {
    if (json == null) {
        return json;
    }
    return {
        
        'size': json['size'],
        'teams': ((json['teams'] as Array<any>).map(TeamRecordFromJSON)),
    };
}

export function TeamSizeGroupToJSON(json: any): TeamSizeGroup {
    return TeamSizeGroupToJSONTyped(json, false);
}

export function TeamSizeGroupToJSONTyped(value?: TeamSizeGroup | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'size': value['size'],
        'teams': ((value['teams'] as Array<any>).map(TeamRecordToJSON)),
    };
}