// src/api/models/TeamStatsResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * radarvan
 * Stats for generals
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TeamSizeGroup } from './TeamSizeGroup';
import {
    TeamSizeGroupFromJSON,
    TeamSizeGroupFromJSONTyped,
    TeamSizeGroupToJSON,
    TeamSizeGroupToJSONTyped,
} from './TeamSizeGroup';

/**
 * 
 * @export
 * @interface TeamStatsResponse
 */
export interface TeamStatsResponse {
    /**
     * 
     * @type {Array<TeamSizeGroup>}
     * @memberof TeamStatsResponse
     */
    groups: Array<TeamSizeGroup>;
}

/**
 * Check if a given object implements the TeamStatsResponse interface.
 */
export function instanceOfTeamStatsResponse(value: object): value is TeamStatsResponse {
    if (!('groups' in value) || value['groups'] === undefined) return false;
    return true;
}

export function TeamStatsResponseFromJSON(json: any): TeamStatsResponse {
    return TeamStatsResponseFromJSONTyped(json, false);
}

export function TeamStatsResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): TeamStatsResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'groups': ((json['groups'] as Array<any>).map(TeamSizeGroupFromJSON)),
    };
}

export function TeamStatsResponseToJSON(json: any): TeamStatsResponse {
    return TeamStatsResponseToJSONTyped(json, false);
}

export function TeamStatsResponseToJSONTyped(value?: TeamStatsResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'groups': ((value['groups'] as Array<any>).map(TeamSizeGroupToJSON)),
    };
}